de-monolith Stage 7: relocate perception_monitor + kinematics_sim + capture to kirra-core; sever the adapter→SDK edge
The final foundational extraction. Lifts the last three lean safety modules out
of the heavy kirra-runtime-sdk into kirra-core so the ROS2 adapter drops the SDK
dependency entirely — and, transitively, the planner loses its last SDK pull.

Moved into kirra-core (bodies code-identical by diff; only the module-path
imports changed — `crate::gateway::kinematics_contract` → `crate::kinematics_contract`,
`kirra_core::PerceptionDerateEvent` → `crate::PerceptionDerateEvent`,
`crate::verifier::FleetPosture` → `crate::FleetPosture`):
  - perception_monitor — Track-C perception derate (KIRRA-OCCY-PMON-001). Lean
    (std::sync + the kinematics-contract talisman + PerceptionDerateEvent).
  - kinematics_sim — bicycle-model forward simulator. Pure physics over the
    contract; pulls only the lightweight `tracing` facade (now an always-on
    kirra-core dep, used by one debug log).
  - capture — learning-loop record builders + bounded-mpsc JSONL writer. Behind a
    new off-by-default `capture` feature so the writer's tokio/serde_json/wire-schema
    deps do NOT reach the consumers that only need the foundation (planner, lane
    map, parko, governor-service): core stays serde-only by default.

SDK re-export shims keep every existing path: `crate::gateway::perception_monitor`,
`crate::kinematics_sim`, `crate::capture` (and their `kirra_runtime_sdk::*`
equivalents) are now thin `pub use kirra_core::…::*;`. The SDK enables
`kirra-core/capture` in its manifest. The verifier command-gateway emit point and
all governor/posture-engine code are unchanged.

Adapter: every `kirra_runtime_sdk::*` import repointed to `kirra_core::*`; the
`kirra-runtime-sdk` path dependency is removed. The `ros2` feature now enables
`kirra-core/capture` (the only capture consumer is the ros2-gated node.rs).

Verification:
  - cargo test --workspace: 1395 passed, 0 failed (58 binaries); relocated
    perception_monitor/kinematics_sim/capture tests run in kirra-core under the
    capture feature.
  - cargo clippy --workspace --all-targets -D warnings: clean.
  - kirra-core rustdoc with -D rustdoc::broken_intra_doc_links: clean.
  - cargo tree: kirra-ros2-adapter and kirra-planner no longer reference the
    kirra-runtime-sdk crate; kirra-core's default-feature tree is serde + tracing
    only (no tokio); tokio + kirra-capture-schema appear only under `capture`.

Drive-by (required to keep the workspace clippy gate green under rust 1.94's new
`cloned_ref_to_slice_refs` lint): kirra-collector/tests/pipeline.rs uses
`std::slice::from_ref` instead of `&[x.clone()]`.
